modules to mmseg
