+++ otx/algorithms/segmentation/adapters/mmseg/datasets/pipelines/__init__.py
@@ -1,0 +1,1 @@
+
