+++ otx/algorithms/segmentation/adapters/mmseg/models/losses/base_weighted_loss.py
@@ -9,7 +9,8 @@ import torch.nn as nn
 from mmseg.core import build_pixel_sampler
 from scipy.special import erfinv
 
-from otx.mpa.modules.models.builder import build_scalar_scheduler
+from otx.algorithms.segmentation.adapters.mmseg.utils.builder import build_scalar_scheduler
+
 
 
 class BaseWeightedLoss(nn.Module, metaclass=ABCMeta):
