+++ otx/algorithms/segmentation/adapters/mmseg/models/schedulers/poly.py
@@ -4,7 +4,7 @@
 
 import numpy as np
 
-from ..builder import SCALAR_SCHEDULERS
+from otx.algorithms.segmentation.adapters.mmseg.utils.builder import SCALAR_SCHEDULERS
 from .base import BaseScalarScheduler
