+++ otx/algorithms/segmentation/adapters/mmseg/models/schedulers/step.py
@@ -6,7 +6,7 @@ from typing import List
 
 import numpy as np
 
-from ..builder import SCALAR_SCHEDULERS
+from otx.algorithms.segmentation.adapters.mmseg.utils.builder import SCALAR_SCHEDULERS
 from .base import BaseScalarScheduler
