+++ otx/algorithms/segmentation/adapters/mmseg/models/segmentors/class_incr_encoder_decoder.py
@@ -9,9 +9,8 @@ from mmseg.utils import get_root_logger
 
 from otx.mpa.modules.utils.task_adapt import map_class_names
 
-from .mix_loss_mixin import MixLossMixin
 from .otx_encoder_decoder import OTXEncoderDecoder
-from .pixel_weights_mixin import PixelWeightsMixin
+from .mixin import MixLossMixin, PixelWeightsMixin
 
 
 @SEGMENTORS.register_module()
