+++ otx/algorithms/segmentation/adapters/mmseg/utils/builder.py
@@ -10,6 +10,23 @@ import torch
 from mmcv.runner import load_checkpoint
 from mmcv.utils import Config, ConfigDict
 
+from mmseg.models.builder import MODELS
+
+SCALAR_SCHEDULERS = MODELS
+
+
+def build_scalar_scheduler(cfg, default_value=None):
+    if cfg is None:
+        if default_value is not None:
+            assert isinstance(default_value, (int, float))
+            cfg = dict(type="ConstantScalarScheduler", scale=float(default_value))
+        else:
+            return None
+    elif isinstance(cfg, (int, float)):
+        cfg = dict(type="ConstantScalarScheduler", scale=float(cfg))
+
+    return SCALAR_SCHEDULERS.build(cfg)
+
 
 def build_segmentor(
     config: Config,
