+++ otx/algorithms/segmentation/adapters/mmseg/utils/data_utils.py
@@ -18,6 +18,7 @@ import json
 import os
 from typing import List, Optional
 
+import torch
 import cv2
 import numpy as np
 import tqdm
@@ -155,6 +156,15 @@ def get_extended_label_names(labels: List[LabelEntity]):
     all_labels = ["background"] + target_labels
     return all_labels
 
+
+def get_valid_label_mask_per_batch(img_metas, num_classes):
+    valid_label_mask_per_batch = []
+    for _, meta in enumerate(img_metas):
+        valid_label_mask = torch.Tensor([1 for _ in range(num_classes)])
+        if "ignored_labels" in meta and meta["ignored_labels"]:
+            valid_label_mask[meta["ignored_labels"]] = 0
+        valid_label_mask_per_batch.append(valid_label_mask)
+    return valid_label_mask_per_batch
 
 @check_input_parameters_type()
 def create_pseudo_masks(ann_file_path: str, data_root_dir: str, mode="FH"):
