--- otx/mpa/modules/models/segmentors/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# flake8: noqa
-from . import class_incr_encoder_decoder, mean_teacher_segmentor, otx_encoder_decoder
