--- otx/mpa/modules/ov/models/mmseg/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# flake8: noqa
-from . import backbones, decode_heads
